requirements.txt
Flask>=2.3.0
requests>=2.31.0
pandas>=2.1.0
pytz>=2024.2
urllib3>=2.0.0

# Authentication & User Management
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.0

# Email Support (for Pro/Elite tiers)
Flask-Mail>=0.9.1

# Environment Variables
python-dotenv>=1.0.0

# Forms & Security
Flask-WTF>=1.2.0
email-validator>=2.1.0

# Rate Limiting
Flask-Limiter>=3.5.0

# Payment Processing (Pro Tier)
stripe>=7.0.0

# Wallet Authentication (Crypto sign-in)
web3>=6.11.0
eth-account>=0.10.0

# Token Generation (for email confirmation)
itsdangerous>=2.1.0

# Future enhancements (Elite tier)
# twilio>=8.0.0  # SMS alerts for Elite tier
# Flask-SocketIO>=5.3.0  # Real-time websocket updates

